Se implemento la navegacion en el front del navbar con react-router-dom
