+++ front/src/components/auth/login.jsx
@@ -1,13 +1,14 @@
 import styles from '../../styles/Login.module.css';
+// import Image from 'next/image';
 
 function Login() {
   return (
     <div className={styles.container}>
       <section className={styles.section}>
-        <img src="/logo.png" alt="Logo" />
+        <img src="/plataforma/front/src/assets/img/logoCuadrado.png" alt="logo" />
       </section>
       <section className={styles.section}>
-      <form>
+      <form className={styles.form}>
       <h2>Iniciar Sesion</h2>
         <label>
           <input className={styles.input} type="text" name="nameUser" placeholder='Nombre Usuario'/>
