--- front/src/components/auth/login.module.css
@@ -1,34 +1,0 @@
-.container{
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    width: 80vh;
-    background-color: rgb(0, 0, 0);
-    border-radius: 5lvh;
-    padding: 20px;
-    color: rgb(255, 255, 255);
-}
-
-.imagen{
-    /* background-color: rgb(255, 255, 255);
-    height: 150px;
-    width: 150px; */
-}
-
-.section{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    border-radius: 15px;
-    padding: 20px;
-    box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);
-}
-
-.input{
-    margin: 10px;
-    padding: 10px;
-    border-radius: 10px;
-    border: 1px solid rgb(0, 0, 0);
-    width: 200px;
-}
