+++ front/src/components/auth/register.jsx
@@ -7,7 +7,7 @@ function Register() {
                 <img src="/logo.png" alt="Logo" />
             </section>
             <section className={styles.section}>
-            <form>
+            <form className={styles.form}>
                 <h2>Register</h2>
                 <label>
                     <input className={styles.input} type="text" name="name" placeholder='Nombres'/>
