+++ front/src/components/home/Home.jsx
@@ -1,0 +1,8 @@
+function Home() {
+    return(
+    <>
+        <h1>Home</h1>
+    </>)
+}
+
+export default Home
