--- front/src/components/navBar/navbar.module.css
@@ -1,22 +1,0 @@
-.navbar{
-    background-color: black;
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    position: absolute;
-    align-items: center;
-}
-
-.navbar .leftLinks{
-    display: flex;
-    gap: 10px;
-}
-
-.navbar .left-links a{
-    color: red;
-    text-decoration: none;
-}
-
-.navbar .rightImage{
-    margin-left: auto; /* Empuja la imagen a la derecha */
-}
