+++ front/src/styles/login.module.css
@@ -1,27 +1,45 @@
+body{
+    margin: 0;
+    padding: 0;
+}
+
 .container{
-    background-color: black;
+    background-color: rgb(0, 0, 0);
     display: flex;
-    justify-content: space-around;
+    width: 100%;
+    height: auto;
+    padding-top: 60px;
+    justify-content: center;
+    align-items: center;
     border: 15px solid black;
 }
 
 .section{
-    width: 50%;
+    width: 40%;
     display: flex;
     justify-content: space-around;
     align-items: center;
+    height: 100%;
 }
 
 .input{
-    width: 80%;
+    width: 100%;
     height: 30px;
     margin: 10px;
     padding: 10px;
     border-radius: 5px;
 }
 
+.form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+
 .button{
     width: 80%;
+    height: 30px;
     margin: 10px;
     padding: 5px;
     border-radius: 5px;
